fix: adding .git folder to the folder should propagate okteto git commit
