+++ cmd/deploy/remote.go
@@ -66,6 +66,9 @@ ENV {{ .RemoteDeployEnvVar }} true
 {{ if ne .ActionNameValue "" }}
 ENV {{ .ActionNameEnvVar }} {{ .ActionNameValue }}
 {{ end }}
+{{ if ne .GitCommitValue "" }}
+ENV {{ .GitCommitEnvVar }} {{ .GitCommitValue }}
+{{ end }}
 
 COPY . /okteto/src
 WORKDIR /okteto/src
@@ -87,6 +90,8 @@ type dockerfileTemplateProperties struct {
 	TokenValue         string
 	ActionNameEnvVar   string
 	ActionNameValue    string
+	GitCommitEnvVar    string
+	GitCommitValue     string
 	RemoteDeployEnvVar string
 	DeployFlags        string
 	RandomInt          int
@@ -201,6 +206,8 @@ func (rd *remoteDeployCommand) createDockerfile(tmpDir string, opts *Options) (s
 		TokenValue:         okteto.Context().Token,
 		ActionNameEnvVar:   model.OktetoActionNameEnvVar,
 		ActionNameValue:    os.Getenv(model.OktetoActionNameEnvVar),
+		GitCommitEnvVar:    model.OktetoGitCommitEnvVar,
+		GitCommitValue:     os.Getenv(model.OktetoGitCommitEnvVar),
 		RemoteDeployEnvVar: constants.OKtetoDeployRemote,
 		RandomInt:          int(randomNumber.Int64()),
 		DeployFlags:        strings.Join(getDeployFlags(opts), " "),
