+++ pkg/repository/repository_test.go
@@ -17,7 +17,6 @@ import (
 	"net/url"
 	"testing"
 
-	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing"
 	"github.com/stretchr/testify/assert"
 )
@@ -63,186 +62,32 @@ func (fs fakeStatus) IsClean() bool {
 }
 
 func TestNewRepo(t *testing.T) {
-	r := NewRepository("https://my-repo/okteto/okteto")
-	assert.Equal(t, "/okteto/okteto", r.url.Path)
-}
-
-func TestIsClean(t *testing.T) {
-	type config struct {
-		repositoryGetter fakeRepositoryGetter
-	}
-	type expected struct {
-		isClean bool
-		err     error
-	}
-	var tests = []struct {
-		name     string
-		config   config
-		expected expected
+	tt := []struct {
+		name            string
+		GitCommit       string
+		expectedControl repositoryInterface
 	}{
 		{
-			name: "dir is not a repository",
-			config: config{
-				repositoryGetter: fakeRepositoryGetter{
-					repository: nil,
-					err:        git.ErrRepositoryNotExists,
-				},
-			},
-			expected: expected{
-				isClean: false,
-				err:     git.ErrRepositoryNotExists,
+			name:      "GitCommit is empty",
+			GitCommit: "",
+			expectedControl: gitRepoController{
+				repoGetter: gitRepositoryGetter{},
 			},
 		},
 		{
-			name: "repository could not access worktree",
-			config: config{
-				repositoryGetter: fakeRepositoryGetter{
-					repository: &fakeRepository{
-						worktree: nil,
-						err:      assert.AnError,
-					},
-				},
-			},
-			expected: expected{
-				isClean: false,
-				err:     assert.AnError,
-			},
-		},
-		{
-			name: "worktree could not access status",
-			config: config{
-				repositoryGetter: fakeRepositoryGetter{
-					repository: &fakeRepository{
-						worktree: &fakeWorktree{
-							status: nil,
-							err:    assert.AnError,
-						},
-						err: nil,
-					},
-				},
-			},
-			expected: expected{
-				isClean: false,
-				err:     assert.AnError,
-			},
-		},
-		{
-			name: "repository is not clean",
-			config: config{
-				repositoryGetter: fakeRepositoryGetter{
-					repository: &fakeRepository{
-						worktree: &fakeWorktree{
-							status: &fakeStatus{
-								isClean: false,
-							},
-							err: nil,
-						},
-						err: nil,
-					},
-				},
-			},
-			expected: expected{
-				isClean: false,
-				err:     nil,
-			},
-		},
-		{
-			name: "repository is clean",
-			config: config{
-				repositoryGetter: fakeRepositoryGetter{
-					repository: &fakeRepository{
-						worktree: &fakeWorktree{
-							status: &fakeStatus{
-								isClean: true,
-							},
-							err: nil,
-						},
-						err: nil,
-					},
-				},
-			},
-			expected: expected{
-				isClean: true,
-				err:     nil,
+			name:      "GitCommit is not empty",
+			GitCommit: "1234567890",
+			expectedControl: oktetoInsideRemoteDeployRepositoryController{
+				gitCommit: "1234567890",
 			},
 		},
 	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			repo := Repository{
-				repositoryGetter: tt.config.repositoryGetter,
-			}
-			isClean, err := repo.IsClean()
-			assert.ErrorIs(t, err, tt.expected.err)
-			assert.Equal(t, tt.expected.isClean, isClean)
-		})
-	}
-}
-
-func TestGetSHA(t *testing.T) {
-	type config struct {
-		repositoryGetter fakeRepositoryGetter
-	}
-	type expected struct {
-		sha string
-		err error
-	}
-	var tests = []struct {
-		name     string
-		config   config
-		expected expected
-	}{
-		{
-			name: "get sha without any problem",
-			config: config{
-				repositoryGetter: fakeRepositoryGetter{
-					repository: &fakeRepository{
-						head: plumbing.NewHashReference("test", plumbing.NewHash("test")),
-					},
-				},
-			},
-			expected: expected{
-				sha: plumbing.NewHash("test").String(),
-				err: nil,
-			},
-		},
-		{
-			name: "error getting repository",
-			config: config{
-				repositoryGetter: fakeRepositoryGetter{
-					repository: nil,
-					err:        assert.AnError,
-				},
-			},
-			expected: expected{
-				sha: "",
-				err: assert.AnError,
-			},
-		},
-		{
-			name: "error getting Head",
-			config: config{
-				repositoryGetter: fakeRepositoryGetter{
-					repository: &fakeRepository{
-						head: nil,
-						err:  assert.AnError,
-					},
-				},
-			},
-			expected: expected{
-				sha: "",
-				err: assert.AnError,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			repo := Repository{
-				repositoryGetter: tt.config.repositoryGetter,
-			}
-			sha, err := repo.GetSHA()
-			assert.ErrorIs(t, err, tt.expected.err)
-			assert.Equal(t, tt.expected.sha, sha)
+	for _, tc := range tt {
+		t.Run(tc.name, func(t *testing.T) {
+			t.Setenv("OKTETO_GIT_COMMIT", tc.GitCommit)
+			r := NewRepository("https://my-repo/okteto/okteto")
+			assert.Equal(t, "/okteto/okteto", r.url.Path)
+			assert.Equal(t, tc.expectedControl, r.repoCtrl)
 		})
 	}
 }
